Add files via upload
Welcome to NLTK, in this we will first discuss with tokenization concept.  a sentence is converted into tokens. eg we have one sentence " Machine learning deals with structured data" in this sentence we have 6 words and each one is one token, like this in the attachment a paragraph is converted into no of tokens